fix(recommend): nine things review found, and the two that stopped a base
Two of them were boot blockers on any base that already exists.

`artifacts.updated_at` was added to the schema with no way in. `migrate`
reads the columns before it applies the file and refuses a base that is
behind, so every install failed to start until somebody hand-ALTERed a
column that nothing reads yet. The doctrine it was obeying — one
statement of what the schema is, recreate to change a column — is right
about columns that change meaning and wrong about this one, which is
additive and defaulted. There is now a named list of exactly such
columns, one entry long. Everything not on it still recreates.

`feedback.enabled`, `associate.enabled` and `pursuit.enabled` were
deleted in favour of `[learn]`, and deserialization drops what it does
not recognise — so a file still carrying them parsed in silence and the
keys did nothing. `[learn]` defaults to on and `pursuit.enabled` shipped
off, which means an upgrade turned recording back on for the operator who
had written `enabled = false` to refuse it, and started writing
model-written artifacts into the index for anyone with `[infer.synthesize]`.
They are refused at startup and named. An alias cannot be written: the
three were independent, and `feedback = true` beside `pursuit = false`
has no single `[learn]` value that means what the file meant.

Then the offer's own.

Forgetting did not reach the situations. `purge_feedback` took the
searches, the questions and the pursuits and left `context_events` — a
device key, an hour, a weekday and a battery reading per page view, kept
400 days — and `context_clusters`, which is those same situations
averaged. The offer went on explaining itself out of data the person had
pressed a button to be rid of. Both tables go now, and the centroids on
the points go with them: they live in the index, where a DELETE cannot
reach, so the handler clears those first while the rows still say which
points carry a set.

Ageing rode `recommends()`, which is `recommend && learn` — so switching
learning off left every 400-day window frozen with everything collected
so far still in it. Turning a thing off is when its data most needs to
leave. The unit is armed by the key that says the feature exists.

The ladder tested one candidate. `best` was an argmax over the full
cosine, which `scope` at weight 10 dominates, and the rung was then
decided by `context_score`, which slices `scope` off. A thin cluster
could win the ranking on the block the gate does not read, fail the gate
on the blocks it does, and drop the page to a random card with an
established Friday afternoon sitting second. Every candidate is tested,
in rank order, and the first that clears a rung is the offer.

`recommended_shown` was written when the offer was computed, not when it
was seen. The fragment races the first keystroke and app.js drops the
answer that loses — so a visit that went straight to typing counted as an
impression that structurally could not be clicked, in the denominator of
the one number the block weights are meant to be fitted against. The
browser confirms it now, through `/ui/context/seen`, and the rung and the
artifact are both checked there: it is a page saying what happened, and a
page can be made to say anything.

And three that are only what they say. `local_hour` truncated the
fractional hour the encoder keeps, making 14:05 and 14:55 the same row;
it is REAL, and the comment no longer claims the sweep reads it, which it
does not — it re-derives all three. `PINNED_TAG` has its doc comment back;
`CTX` had been inserted underneath it.

+++ src/jobs/context.rs
@@ -546,7 +546,7 @@ mod tests {
                 at,
                 bundle: raw.clone(),
                 device_key: device_key(&parse_bundle(&raw)),
-                local_hour: Some(t.hour as i64),
+                local_hour: Some(t.hour as f64),
                 weekday: Some(t.weekday as i64),
                 tz: bundle.tz.clone(),
             })
